+++ packages/plugins/src/plugins/bonfida/constants.ts
@@ -1,4 +1,5 @@
 import { PublicKey } from '@solana/web3.js';
+import { AirdropStatics } from '../../AirdropFetcher';
 
 export const platformId = 'sns';
 export const fidaMint = 'EchesyfXePKdLtoiZSL8pBe8Myagyy8ZRqsACNCFGnvp';
@@ -13,3 +14,16 @@ export const airdropPid = new PublicKey(
 export const airdropApi =
   'https://jito-distributor-api.n38t0uklh84us.ap-southeast-1.cs.amazonlightsail.com/proof/';
 export const fidaDecimals = 6;
+
+export const snsMint = 'SNS8DJbHc34nKySHVhLGMUUE72ho6igvJaxtq9T3cX3';
+
+export const snsAirdropStatics: AirdropStatics = {
+  emitterLink: 'https://sns.id/',
+  emitterName: 'sns',
+  id: 'sns-airdrop-sns',
+  claimLink: 'https://airdrop.sns.id/',
+  claimStart: 1747137600000,
+  claimEnd: 1754913600000,
+  image:
+    'https://sonarwatch.github.io/portfolio/assets/images/platforms/debridge.webp',
+};