+++ packages/tx-parser/src/services.ts
@@ -108,6 +108,7 @@ import * as rain from './services/rain';
 import * as ratex from './services/ratex';
 import * as raydium from './services/raydium';
 import * as realms from './services/realms';
+import * as relay from './services/relay';
 import * as runemine from './services/runemine';
 import * as sanctum from './services/sanctum';
 import * as sandglass from './services/sandglass';
@@ -120,6 +121,7 @@ import * as sns from './services/sns';
 import * as solana from './services/solana';
 import * as solanaid from './services/solanaid';
 import * as solayer from './services/solayer';
+import * as solcasino from './services/solcasino';
 import * as solincinerator from './services/solincinerator';
 import * as solsea from './services/solsea';
 import * as sonic from './services/sonic';
@@ -258,6 +260,7 @@ export const services: ServiceDefinition[] = [
   ratex,
   raydium,
   realms,
+  relay,
   runemine,
   sanctum,
   sandglass,
@@ -270,6 +273,7 @@ export const services: ServiceDefinition[] = [
   solana,
   solanaid,
   solayer,
+  solcasino,
   solincinerator,
   solsea,
   sonic,